Solved plate cheese
Allowed sides again. Made perform update check if there is one or two potatoes and hid/show the side if it needs to. Fixed the plateup chef issue by using a dummy gameobject for the jacket potato as chef treats the ordering item as if the potato is not there for some reason